requirements.txt
click==6.7
numpy==1.13.3
pandas==0.21.1
PyGithub==1.35
PyJWT==1.5.3
python-dateutil==2.6.1
pytz==2017.3
six==1.11.0
